+++ R/gdoc2rmd.R
@@ -2,20 +2,16 @@
 #'
 #' Use Articul8 Add-ons from Google docs to build Rticles
 #'
-#' @param file Zip file path from Articul8 exported in  md format
-#' @param export path to export the files. Default file directory
-#' @param prefix_fig Prefix for the name of the figure
-#' @param prefix_tab Prefix for the name of the table
+#' @param file Zip file path from Articul8 exported in md format (path)
+#' @param export Path to export the files (path: "docs")
 #'
-#' @return folder
+#' @return path
 #' 
 #' @export
 #' 
 
 gdoc2rmd <- function(file
-                     , export = "files"
-                     , prefix_fig = "Figure"
-                     , prefix_tab = "Table"
+                     , export = "docs"
                      ){
   
   zip <- file %>% 
@@ -26,11 +22,12 @@ gdoc2rmd <- function(file
     readLines() %>% 
     tibble::enframe() %>%
     dplyr::rowwise() %>%
-    dplyr::mutate(value = figure2rmd(.data$value, path = export, prefix = prefix_fig)) %>% 
-    dplyr::mutate(value = table2rmd(.data$value, prefix = prefix_tab)) %>% 
+    dplyr::mutate(value = figure2rmd(.data$value, path = export)) %>% 
+    dplyr::mutate(value = table2rmd(.data$value)) %>%
+    dplyr::mutate(value = gsub("```Unknown element type at this position: UNSUPPORTED```", "\n", .data$value)) %>% 
     dplyr::select(.data$value) %>% 
     tibble::deframe() %>% 
-    writeLines(con = paste0(export,"/_doc.Rmd"))
+    writeLines(con = file.path(export, "_doc.Rmd") %>% gsub("\\\\", "\\/", .))
   
   doc <- list.files(path = export, pattern = "_doc", full.names = T)